requirements.txt
# Core Framework
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

# LangChain & LangGraph
langchain==1.2.3
langchain-core==1.2.7
langchain-community==0.4.1
langchain-mistralai==1.1.1
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
langchain-groq==1.1.1
langchain-postgres==0.0.16
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint
langgraph-checkpoint-postgres==3.0.3
langsmith==0.4.13

# Database
psycopg2-binary==2.9.11
psycopg==3.3.2
pgvector==0.3.6
SQLAlchemy==2.0.43

# Document Processing
pypdf==6.0.0

# HTTP Client
httpx==0.28.1
requests==2.32.5

# Utilities
tenacity==9.1.2
tiktoken==0.11.0
PyYAML==6.0.2
numpy==2.3.2  